+++ src/interactivity/win32/UiaTextRange.cpp
@@ -723,9 +723,9 @@ IFACEMETHODIMP UiaTextRange::GetText(_In_ int maxLength, _Out_ BSTR* pRetVal)
                 const ROW& row = pTextBuffer->GetRowAtIndex(rowIndex);
                 if (row.CharRow.ContainsText())
                 {
-                    const int rowRight = row.CharRow.MeasureRight();
-                    int startIndex = 0;
-                    int endIndex = rowRight;
+                    const size_t rowRight = row.CharRow.MeasureRight();
+                    size_t startIndex = 0;
+                    size_t endIndex = rowRight;
                     if (currentScreenInfoRow == startScreenInfoRow)
                     {
                         startIndex = startColumn;
@@ -733,7 +733,7 @@ IFACEMETHODIMP UiaTextRange::GetText(_In_ int maxLength, _Out_ BSTR* pRetVal)
                     if (currentScreenInfoRow == endScreenInfoRow)
                     {
                         // prevent the end from going past the last non-whitespace char in the row
-                        endIndex = min(static_cast<int>(endColumn + 1), rowRight);
+                        endIndex = min(endColumn + 1, rowRight);
                     }
 
                     // if startIndex >= endIndex then _start is
